Store version numbers as arrays of ints instead of text
